requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.3.0
Authlib==1.3.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dataclasses-json==0.6.7
datasets==3.2.0
dill==0.3.8
distro==1.9.0
docarray==0.40.0
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
extra-streamlit-components==0.1.71
filelock==3.17.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-health-checking==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-oauth==0.15.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.13
langchain-community==0.3.12
langchain-core==0.3.27
langchain-ollama==0.2.2
langchain-openai==0.2.13
langchain-text-splitters==0.3.4
langchain-weaviate==0.0.3
langdetect==1.0.9
langsmith==0.2.4
Levenshtein==0.26.1
lightning==2.5.0.post0
lightning-utilities==0.12.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
narwhals==1.18.4
networkx==3.2.1
nltk==3.9.1
nougat==0.3.3
nougat-ocr==0.1.17
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.47
ollama==0.4.4
openai==1.58.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.12
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.3
pymupdf4llm==0.0.17
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dotenv==1.0.1
python-iso639==2024.10.22
python-Levenshtein==0.26.1
python-magic==0.4.27
python-oxmsg==0.0.1
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scipy==1.13.1
sconf==0.2.5
sentencepiece==0.2.0
simsimd==6.2.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.41.1
streamlit-chat==0.1.1
streamlit-google-auth==1.1.8
streamlit-oauth==0.1.14
stringzilla==3.11.3
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
timm==0.5.4
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
tzdata==2024.2
unstructured==0.16.11
unstructured-client==0.28.1
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
weaviate-client==4.10.2
webencodings==0.5.1
websockets==14.2
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
